+++ TabLayout/tests/TabLayout-specs.js
@@ -168,10 +168,10 @@ describe('TabLayout specs', () => {
 		expect(actual).toMatchObject(expected);
 	});
 
-	test('should call \'onSelect\' with \'onSelect\' type when pressing \'Enter\' on a tab', () => {
+	test('should call \'onSelect\' with \'onSelect\' type when pressing \'Enter\' on a tab while \'vertial\' and \'ltr\'', () => {
 		const spy = jest.fn();
 		render(
-			<TabLayout onSelect={spy} orientation="vertical">
+			<TabLayout onSelect={spy} orientation="vertical" rtl={false}>
 				<Tab icon="home" title="Home">
 					<div>Home</div>
 				</Tab>
@@ -182,6 +182,106 @@ describe('TabLayout specs', () => {
 		);
 
 		const tab = screen.getAllByTestId('tab')[1];
+		enterKeyDown(tab);
+		enterKeyUp(tab);
+
+		const expected = {type: 'onSelect'};
+		const actual = spy.mock.calls.length && spy.mock.calls[0][0];
+
+		expect(actual).toMatchObject(expected);
+	});
+
+	test('should call \'onSelect\' with \'onSelect\' type when pressing \'Enter\' on a tab while \'vertial\' and \'rtl\'', () => {
+		const spy = jest.fn();
+		render(
+			<TabLayout onSelect={spy} orientation="vertical" rtl>
+				<Tab data-testid="tab" title="Home">
+					<div>Home</div>
+				</Tab>
+			</TabLayout>
+		);
+
+		const tab = screen.getAllByTestId('tab')[0];
+		enterKeyDown(tab);
+		enterKeyUp(tab);
+
+		const expected = {type: 'onSelect'};
+		const actual = spy.mock.calls.length && spy.mock.calls[0][0];
+
+		expect(actual).toMatchObject(expected);
+	});
+
+	test('should call \'onSelect\' with \'onSelect\' type when pressing \'Enter\' on a tab while \'end\' and \'ltr\'', () => {
+		const spy = jest.fn();
+		render(
+			<TabLayout onSelect={spy} anchorTo="end" rtl={false}>
+				<Tab data-testid="tab" title="Home">
+					<div>Home</div>
+				</Tab>
+			</TabLayout>
+		);
+
+		const tab = screen.getAllByTestId('tab')[0];
+		enterKeyDown(tab);
+		enterKeyUp(tab);
+
+		const expected = {type: 'onSelect'};
+		const actual = spy.mock.calls.length && spy.mock.calls[0][0];
+
+		expect(actual).toMatchObject(expected);
+	});
+
+	test('should call \'onSelect\' with \'onSelect\' type when pressing \'Enter\' on a tab while \'end\' and \'rtl\'', () => {
+		const spy = jest.fn();
+		render(
+			<TabLayout onSelect={spy} anchorTo="end" rtl>
+				<Tab data-testid="tab" title="Home">
+					<div>Home</div>
+				</Tab>
+			</TabLayout>
+		);
+
+		const tab = screen.getAllByTestId('tab')[0];
+		enterKeyDown(tab);
+		enterKeyUp(tab);
+
+		const expected = {type: 'onSelect'};
+		const actual = spy.mock.calls.length && spy.mock.calls[0][0];
+
+		expect(actual).toMatchObject(expected);
+	});
+
+	test('should call \'onSelect\' with \'onSelect\' type when pressing \'Enter\' on a tab while \'right\'', () => {
+		const spy = jest.fn();
+		render(
+			<TabLayout onSelect={spy} anchorTo="right">
+				<Tab data-testid="tab" title="Home">
+					<div>Home</div>
+				</Tab>
+			</TabLayout>
+		);
+
+		const tab = screen.getAllByTestId('tab')[0];
+		enterKeyDown(tab);
+		enterKeyUp(tab);
+
+		const expected = {type: 'onSelect'};
+		const actual = spy.mock.calls.length && spy.mock.calls[0][0];
+
+		expect(actual).toMatchObject(expected);
+	});
+
+	test('should call \'onSelect\' with \'onSelect\' type when pressing \'Enter\' on a tab while \'horizontal\'', () => {
+		const spy = jest.fn();
+		render(
+			<TabLayout onSelect={spy} orientation="horizontal">
+				<Tab data-testid="tab" title="Home">
+					<div>Home</div>
+				</Tab>
+			</TabLayout>
+		);
+
+		const tab = screen.getAllByTestId('tab')[0];
 		enterKeyDown(tab);
 		enterKeyUp(tab);